requirements.txt
typer[all]>=0.9.0
pyyaml>=6.0
numpy>=1.24.0
openai>=2.8.1
python-dotenv>=1.0.0
xai-sdk>=1.4.1
anthropic>=0.75.0
google-genai>=1.55.0

